service/metricsroute_test.go
// Copyright 2021 Mia srl
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package service

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/gorilla/mux"
	"github.com/prometheus/client_golang/prometheus"
	"github.com/stretchr/testify/require"
)

func TestMetricsRoute(t *testing.T) {
	t.Run("exposes metrics route", func(t *testing.T) {
		router := mux.NewRouter()
		registry := prometheus.NewRegistry()
		metricsRoute(router, registry)

		req := httptest.NewRequest(http.MethodGet, metricsRoutePath, nil)
		w := httptest.NewRecorder()

		router.ServeHTTP(w, req)

		require.Equal(t, http.StatusOK, w.Result().StatusCode)
	})

	t.Run("without registry not exposes route", func(t *testing.T) {
		router := mux.NewRouter()
		metricsRoute(router, nil)

		req := httptest.NewRequest(http.MethodGet, metricsRoutePath, nil)
		w := httptest.NewRecorder()

		router.ServeHTTP(w, req)

		require.Equal(t, http.StatusNotFound, w.Result().StatusCode)
	})
}
